Add option to use an installation prefix
If CMAKE_INSTALL_PREFIX is defined and not empty, the extension will be
installed there instead of the paths from pgconfig.

Also, a helper script is added to create symlinks in the pgconfig
directories to the files in the prefix directories so development
setups can install new versions of pg-gvm without root permissions.