Fix next_time_ical and add reference_time param
The next_time_ical function now has a new parameter for the
reference time, which should be the current time.
This allows using a consistent reference time throughout queries that
take a longer time to run.
The function now no longer returns the current/reference time as the
next time when it is the same as a recurrence time, which could cause
unwanted effects like scheduled tasks being started twice for the same
scheduled time.