change: Use up to date tooling for installing dependencies
Use the same tooling as in gvmd, gvm-libs and other C based repos to
install build and runtime dependencies. This makes it easier to maintain
them. Also use gvm-libs stable as base image and install postgres
generic dependencies for not having to maintain versioned postgres
packages in the install files.